Rename cumulus to slurm. Rename filenames Rename setup to initializeEnvironment Bugfix: use cpp/build/mount instead of bin/mount Bugfix: uuid OSD* to distinct New: mount on each cumu-n** instead on cumulus New: changed some echos to debug outputs incl. new debug output function New: flag to use given configs or prepare configs New: summary output at the end Adjusted readme; explained specific OSD count

+++ xtreemfs_slurm_rstart.sh
@@ -5,10 +5,10 @@
 # Date: 18.05.2015
 #
 # Start script for the Xtreemfs Server.
-# "RStart", cause it's called remote on a specific cumulus node.
+# "rstart", cause it's called remote on a specific slurm node.
 # 
 # Call: 
-# 	./xtreemfsCumulusRStart.sh /path/to/env.sh (DIR|MRC|OSD) (DIR|MRC|OSD*)
+# 	./xtreemfs_slurm_rstart.sh /path/to/env.sh (DIR|MRC|OSD) (DIR|MRC|OSD*)
 #
 # Parameter:
 # 	$1 path to source file (env.sh)
@@ -36,31 +36,35 @@ source $SOURCE_FILE
 # Start xtreemfs server, redirect output into a log and save process id into pid file
 function startServer() {
   
-  SERVER_PID=$(substitude_name "$PID_FILENAME_GENERIC" "$SERVER_NAME") 
-  SERVER_LOG=$(substitude_name "$LOG_FILENAME_GENERIC" "$SERVER_NAME")
-  SERVER_CONFIG=$(substitude_name "$CONFIG_FILENAME_GENERIC" "$SERVER_NAME")
+  SERVER_PID=$(substitudeName "$PID_FILENAME_GENERIC" "$SERVER_NAME") 
+  SERVER_LOG=$(substitudeName "$LOG_FILENAME_GENERIC" "$SERVER_NAME")
+  SERVER_CONFIG=$(substitudeName "$CONFIG_FILENAME_GENERIC" "$SERVER_NAME")
   
-  CURRENT_JOB_FOLDER=$(substitude_jobid "$CURRENT_JOB_FOLDER_GENERIC")
-  CURRENT_LOCAL_FOLDER=$(substitude_jobid "$LOCAL_DIR_GENERIC")
+  CURRENT_JOB_FOLDER=$(substitudeJobID "$CURRENT_JOB_FOLDER_GENERIC")
+  CURRENT_LOCAL_FOLDER=$(substitudeJobID "$LOCAL_DIR_GENERIC")
   mkdir -p $CURRENT_LOCAL_FOLDER
   
   SERVER_TYPE_SMALL="${SERVER_TYPE,,}"
   
   #copy config file to local
-  
-  cp "$CURRENT_JOB_FOLDER/config/$SERVER_CONFIG" "$CURRENT_LOCAL_FOLDER/$SERVER_CONFIG"
+  if [[ -e "$CURRENT_JOB_FOLDER/config/$SERVER_CONFIG" ]]; then
+    cp "$CURRENT_JOB_FOLDER/config/$SERVER_CONFIG" "$CURRENT_LOCAL_FOLDER/$SERVER_CONFIG"
+  else
+    echo "Couldn't find the config file: $CURRENT_JOB_FOLDER/config/$SERVER_CONFIG"
+    return 1
+  fi
   
   #start server
-  echo -n "Starting XtreemFS $SERVER_TYPE Server on $(hostname): $SERVER_NAME ..."
+  outputDebug -n "Starting XtreemFS $SERVER_TYPE Server on $(hostname): $SERVER_NAME ..."
   $JAVA_HOME/bin/java -ea -cp ${JAVA_CLASSPATH} "org.xtreemfs.${SERVER_TYPE_SMALL}.${SERVER_TYPE}" "$CURRENT_LOCAL_FOLDER/$SERVER_CONFIG" >> "$CURRENT_LOCAL_FOLDER/$SERVER_LOG" 2>&1 &
   PROCPID="$!"
   echo "$PROCPID" > "$CURRENT_LOCAL_FOLDER/$SERVER_PID"
   sleep 1s
   
   if [[ -e "/proc/$PROCPID" ]]; then
-   echo "success"
+   outputDebug "success"
   else
-   echo "failed"
+   outputDebug "failed"
    return 1
   fi
